@@ -12,67 +12,68 @@
     "e2e": "ng e2e"
   },
   "engines": {
-    "node": "10.0.0",
+    "node": "10.9.0",
     "npm": "5.3.0"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.7",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.7",
-    "@angular/compiler": "^6.1.7",
-    "@angular/compiler-cli": "^6.1.7",
-    "@angular/core": "^6.1.7",
-    "@angular/fire": "^5.0.0",
-    "@angular/flex-layout": "^6.0.0-beta.18",
-    "@angular/forms": "^6.1.7",
-    "@angular/http": "^6.1.7",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^6.1.7",
-    "@angular/platform-browser-dynamic": "^6.1.7",
-    "@angular/router": "^6.1.7",
-    "@ngrx/effects": "6.1.0",
-    "@ngrx/router-store": "6.1.0",
-    "@ngrx/store": "6.1.0",
-    "@ngrx/store-devtools": "6.1.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.0.0",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "@ngrx/effects": "7.4.0",
+    "@ngrx/router-store": "7.4.0",
+    "@ngrx/store": "7.4.0",
+    "@ngrx/store-devtools": "7.4.0",
     "ag-grid": "^18.1.2",
-    "ag-grid-angular": "^18.1.0",
-    "core-js": "^2.5.7",
-    "dotenv": "^5.0.1",
-    "express": "^4.16.3",
-    "firebase": "^5.4.2",
+    "ag-grid-angular": "^21.0.0",
+    "ag-grid-community": "^21.0.0",
+    "core-js": "^3.1.3",
+    "dotenv": "^8.0.0",
+    "express": "^4.17.1",
+    "firebase": "^6.1.0",
     "hammerjs": "^2.0.8",
     "jwt-decode": "^2.2.0",
-    "moment": "^2.22.2",
-    "ng2-search-filter": "^0.4.7",
+    "moment": "^2.24.0",
+    "ng2-search-filter": "^0.5.1",
     "ngrx-actions": "^4.0.0",
     "ngrx-store-freeze": "^0.2.4",
-    "ngx-order-pipe": "^2.0.1",
-    "rxjs": "^6.3.2",
-    "typescript": "~2.9.2",
-    "yargs": "^11.1.0",
-    "zone.js": "^0.8.19"
+    "ngx-order-pipe": "^2.0.2",
+    "rxjs": "^6.5.2",
+    "typescript": "~3.4.5",
+    "yargs": "^13.2.4",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular/cli": "^6.2.1",
-    "@angular/language-service": "^6.1.7",
+    "@angular-devkit/build-angular": "^0.800.1",
+    "@angular/cli": "^8.0.1",
+    "@angular/language-service": "^8.0.0",
     "@angularclass/hmr": "^2.1.3",
-    "@types/jasmine": "^2.8.8",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.117",
-    "codelyzer": "^4.4.4",
+    "@types/jasmine": "^3.3.13",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "^12.0.4",
+    "codelyzer": "^5.1.0",
     "enhanced-resolve": "^4.1.0",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.5",
+    "karma": "^4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.9.3",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1"
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "node-sass": "^4.12.0",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.2.0",
+    "tslint": "~5.17.0"
   }
 }
