@@ -1,16 +1,18 @@
-import {Component} from '@angular/core';
+import {Component, OnDestroy} from '@angular/core';
 import {ICellRendererAngularComp} from 'ag-grid-angular';
 import {getTour} from '../../store/tour/tour.reducer';
 import {IAppState} from '../../store/store';
 import {Store} from '@ngrx/store';
 import {ITour} from '../../models/tour.model';
+import {takeUntil} from 'rxjs/operators';
+import {Subject} from 'rxjs';
 
 @Component({
   selector: 'app-hastourendedclass',
   templateUrl: './hastourendedclass.component.html',
   styleUrls: ['./hastourendedclass.component.scss']
 })
-export class HastourendedclassComponent implements ICellRendererAngularComp {
+export class HastourendedclassComponent implements ICellRendererAngularComp, OnDestroy {
 
   constructor(private store: Store<IAppState>) {
   }
@@ -18,11 +20,12 @@ export class HastourendedclassComponent implements ICellRendererAngularComp {
   private params: any;
   public tour: ITour;
   private content: string;
+  unsubscribe = new Subject<void>();
 
   // called on init
   agInit(params: any): void {
     this.params = params;
-    this.store.select(getTour).subscribe(tour => {
+    this.store.select(getTour).pipe(takeUntil(this.unsubscribe)).subscribe(tour => {
       this.tour = tour;
     });
   }
@@ -34,5 +37,10 @@ export class HastourendedclassComponent implements ICellRendererAngularComp {
   public value(): string {
     this.content = this.params.value;
     return this.content;
+  }
+
+  ngOnDestroy() {
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
   }
 }
