@@ -1,22 +1,24 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {AuthService} from './services/auth.service';
 import {select, Store} from '@ngrx/store';
 import {IAppState} from './store/store';
 import * as fromParticipanttable from './store/participanttable/participanttable.actions';
 import * as fromTour from './store/tour/tour.actions';
 import {getTour, getTours, isRegistrationOpen} from './store/tour/tour.reducer';
-import {Observable} from 'rxjs';
+import {Observable, Subject} from 'rxjs';
 import {ITour} from './models/tour.model';
 import {getLastUpdated} from './store/participanttable/participanttable.reducer';
 import {MatSnackBar} from '@angular/material';
 import * as fromParticipantForm from './store/participantform/participantform.actions';
+import {getParticipant} from './store/participant/participant.reducer';
+import {takeUntil} from 'rxjs/operators';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent implements OnInit {
+export class AppComponent implements OnInit, OnDestroy {
   constructor(private store: Store<IAppState>, public authService: AuthService, public snackBar: MatSnackBar) {
   }
 
@@ -30,6 +32,7 @@ export class AppComponent implements OnInit {
   lastUpdated$: Observable<any>;
   isRegistrationOpen$: Observable<boolean>;
   lastUpdated: any;
+  unsubscribe = new Subject<void>();
 
   ngOnInit() {
     this.setViewHeight();
@@ -41,25 +44,25 @@ export class AppComponent implements OnInit {
     this.store.dispatch(new fromTour.FetchTour());
 
     this.isRegistrationOpen$ = this.store.pipe(select(isRegistrationOpen));
-
-    this.tours$ = this.store.pipe(select(getTours));
-    this.tours$.subscribe(tours => {
-      if (tours && tours.length > 0) {
-        this.selectedTour = tours.find(tour => tour.isActive);
+    this.store.pipe(select(getParticipant)).pipe(takeUntil(this.unsubscribe)).subscribe(participant => {
+      if (participant) {
+        this.displayName = participant.displayName;
       }
     });
 
+    this.tours$ = this.store.pipe(select(getTours));
+
     this.tour$ = this.store.pipe(select(getTour));
-    this.tour$.subscribe(tour => {
+    this.tour$.pipe(takeUntil(this.unsubscribe)).subscribe(tour => {
       if (tour && tour.id) {
-        console.log('tour is gewijzigd naar: ' + tour.id);
+        this.selectedTour = tour;
         this.store.dispatch(new fromParticipanttable.FetchParticipanttable(tour.id));
         this.store.dispatch(new fromParticipanttable.FetchLastUpdated(tour.id));
       }
     });
     this.lastUpdated$ = this.store.pipe(select(getLastUpdated));
 
-    this.lastUpdated$.subscribe(response => {
+    this.lastUpdated$.pipe(takeUntil(this.unsubscribe)).subscribe(response => {
       if (response &&
         this.lastUpdated && response.tour === this.lastUpdated.tour &&
         response.lastUpdated !== this.lastUpdated.lastUpdated) {
@@ -76,6 +79,10 @@ export class AppComponent implements OnInit {
     this.store.dispatch(new fromTour.FetchTourById(this.selectedTour.id));
   }
 
+  selectTour(tour: ITour) {
+    this.store.dispatch(new fromTour.FetchTourById(tour.id));
+  }
+
   logout() {
     this.authService.logout();
     this.store.dispatch(new fromParticipantForm.ClearParticipantform());
@@ -84,5 +91,10 @@ export class AppComponent implements OnInit {
   setViewHeight() {
     const vh = window.innerHeight;
     document.documentElement.style.setProperty('--vh', `${vh}px`);
+  }
+
+  ngOnDestroy() {
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
   }
 }
