@@ -1,4 +1,4 @@
-import {Component, Inject, OnInit} from '@angular/core';
+import {Component, Inject, OnDestroy, OnInit} from '@angular/core';
 import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material';
 import {FormControl, FormGroup, Validators} from '@angular/forms';
 import {MyErrorStateMatcher} from '../login/login.component';
@@ -6,23 +6,28 @@ import {EtappeService} from '../../services/etappe.service';
 import {IEtappe} from '../../models/etappe.model';
 import {IAppState} from '../../store/store';
 import {Store} from '@ngrx/store';
-import * as fromTour from '../../store/tour/tour.actions';
 import {TourService} from '../../services/tour.service';
 import {getEtappes} from '../../store/etappe/etappe.reducer';
+import {Subject} from 'rxjs';
+import {takeUntil} from 'rxjs/operators';
 
 @Component({
   selector: 'app-edittourriderdialog',
   templateUrl: './edittourriderdialog.component.html',
   styleUrls: ['./edittourriderdialog.component.scss']
 })
-export class EdittourriderdialogComponent implements OnInit {
+export class EdittourriderdialogComponent implements OnInit, OnDestroy {
 
-  constructor(private store: Store<IAppState>, private etappeService: EtappeService, private tourriderservice: TourService, public dialogRef: MatDialogRef<EdittourriderdialogComponent>,
+  constructor(private store: Store<IAppState>,
+              private etappeService: EtappeService,
+              private tourriderservice: TourService,
+              public dialogRef: MatDialogRef<EdittourriderdialogComponent>,
               @Inject(MAT_DIALOG_DATA) public data: any) {
   }
 
   etappes: IEtappe[];
   selectedEtappe: IEtappe;
+  unsubscribe = new Subject<void>();
 
   riderForm = new FormGroup({
     riderWaardeFormControl: new FormControl('', [
@@ -38,7 +43,8 @@ export class EdittourriderdialogComponent implements OnInit {
   ngOnInit() {
     // todo store in store
     // todo get tour?
-    this.store.select(getEtappes).subscribe(response => this.etappes = response.sort((a, b) => a.etappeNumber - b.etappeNumber));
+    this.store.select(getEtappes).pipe(takeUntil(this.unsubscribe))
+      .subscribe(response => this.etappes = response.sort((a, b) => a.etappeNumber - b.etappeNumber));
     this.selectedEtappe = this.data.latestEtappe;
   }
 
@@ -52,8 +58,12 @@ export class EdittourriderdialogComponent implements OnInit {
       rider: data.rider,
       latestEtappe: this.selectedEtappe
     }).subscribe(response => {
-     console.log('saveriderToTeam response: ' + response);
+      console.log('saveriderToTeam response: ' + response);
     });
   }
 
+  ngOnDestroy() {
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
+  }
 }
