@@ -1,20 +1,21 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {AuthService} from '../../services/auth.service';
 import {IAppState} from '../../store/store';
 import {select, Store} from '@ngrx/store';
 import {getLastUpdated, getNumberOne} from '../../store/participanttable/participanttable.reducer';
-import {Observable} from 'rxjs';
+import {Observable, Subject} from 'rxjs';
 import {IParticipanttable} from '../../models/participanttable.model';
 import {getTour, isRegistrationOpen} from '../../store/tour/tour.reducer';
 import {ITour} from '../../models/tour.model';
 import * as moment from 'moment';
+import {takeUntil} from 'rxjs/operators';
 
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss']
 })
-export class HomeComponent implements OnInit {
+export class HomeComponent implements OnInit, OnDestroy {
 
   lastUpdated$: Observable<any>;
   lastUpdated: string;
@@ -22,6 +23,7 @@ export class HomeComponent implements OnInit {
   first$: Observable<IParticipanttable>;
   tour$: Observable<ITour>;
   deadline: Date;
+  unsubscribe = new Subject<void>();
 
   constructor(public authService: AuthService, public store: Store<IAppState>) {
     moment.locale('nl');
@@ -31,7 +33,7 @@ export class HomeComponent implements OnInit {
 
 
     this.lastUpdated$ = this.store.pipe(select(getLastUpdated));
-    this.lastUpdated$.subscribe(lastupdated => {
+    this.lastUpdated$.pipe(takeUntil(this.unsubscribe)).subscribe(lastupdated => {
       if (lastupdated) {
         this.lastUpdated = moment(lastupdated.lastUpdated).fromNow();
       }
@@ -41,11 +43,15 @@ export class HomeComponent implements OnInit {
 
     this.tour$ = this.store.select(getTour);
 
-    this.tour$.subscribe(tour => {
+    this.tour$.pipe(takeUntil(this.unsubscribe)).subscribe(tour => {
       if (tour && tour.deadline) {
         this.deadline = tour.deadline;
       }
     });
   }
 
+  ngOnDestroy() {
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
+  }
 }
