@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {AuthService} from '../../services/auth.service';
 import {Router} from '@angular/router';
 import {FormControl, FormGroup, FormGroupDirective, NgForm, Validators} from '@angular/forms';
@@ -7,19 +7,22 @@ import {ParticipantService} from '../../services/participant.service';
 import {IAppState} from '../../store/store';
 import * as fromParticipantForm from '../../store/participantform/participantform.actions';
 import {Store} from '@ngrx/store';
+import {takeUntil} from 'rxjs/operators';
+import {Subject} from 'rxjs';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss']
 })
-export class LoginComponent implements OnInit {
+export class LoginComponent implements OnInit, OnDestroy {
   user = {
     email: '',
     password: '',
     displayName: '',
     teamName: '',
   };
+  unsubscribe = new Subject<void>();
 
   constructor(public authService: AuthService,
               public participantService: ParticipantService,
@@ -33,11 +36,12 @@ export class LoginComponent implements OnInit {
       Validators.required,
       Validators.email,
     ]),
-    displayNameFormControl: new FormControl('', [
+    displayName: new FormControl('', [
       Validators.required,
     ]),
-    displayTeamNameFormControl: new FormControl('', [
+    teamName: new FormControl('', [
       Validators.required,
+      Validators.maxLength(25)
     ]),
     passwordFormControl: new FormControl('', [
       Validators.required,
@@ -84,7 +88,7 @@ export class LoginComponent implements OnInit {
               displayName: this.user.displayName,
               teamName: this.user.teamName,
               email: this.user.email
-            }).subscribe(response => {
+            }).pipe(takeUntil(this.unsubscribe)).subscribe(response => {
               console.log('user opgeslagen in database');
             });
             this.store.dispatch(new fromParticipantForm.ClearParticipantform());
@@ -106,6 +110,11 @@ export class LoginComponent implements OnInit {
   activateResetPassword(isTrue: boolean) {
     this.wachtwoordvergeten = isTrue;
   }
+
+  ngOnDestroy() {
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
+  }
 }
 
 /** Error when invalid control is dirty, touched, or submitted. */
@@ -114,4 +123,5 @@ export class MyErrorStateMatcher implements ErrorStateMatcher {
     const isSubmitted = form && form.submitted;
     return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
   }
+
 }
