@@ -1,23 +1,24 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {GridOptions} from 'ag-grid';
 import {MatDialog} from '@angular/material';
 import {Router} from '@angular/router';
 import {select, Store} from '@ngrx/store';
 import {IAppState} from '../../store/store';
 import {getLastUpdated, getParticipanttable} from '../../store/participanttable/participanttable.reducer';
-import {Observable} from 'rxjs';
+import {Observable, Subject} from 'rxjs';
 import {IParticipanttable} from '../../models/participanttable.model';
 import {getTour} from '../../store/tour/tour.reducer';
 import {ITour} from '../../models/tour.model';
 import {HastourendedclassComponent} from '../../aggridcomponents/hastourendedclass/hastourendedclass.component';
 import * as moment from 'moment';
+import {takeUntil} from 'rxjs/operators';
 
 @Component({
   selector: 'app-participanttable',
   templateUrl: './participanttable.component.html',
   styleUrls: ['./participanttable.component.scss']
 })
-export class ParticipanttableComponent implements OnInit {
+export class ParticipanttableComponent implements OnInit, OnDestroy {
   private gridApi;
   private gridColumnApi;
   private agColumns;
@@ -29,6 +30,7 @@ export class ParticipanttableComponent implements OnInit {
   data: any;
   participantstable$: Observable<IParticipanttable[]>;
   lastUpdated$: Observable<any>;
+  unsubscribe = new Subject<void>();
 
   public gridOptions: GridOptions;
   public lastUpdated: string;
@@ -38,7 +40,7 @@ export class ParticipanttableComponent implements OnInit {
 
     this.agColumns = [
       {headerName: '#', valueGetter: this.formatPosition, minWidth: 75, maxWidth: 75},
-      {headerName: 'Naam', field: 'displayName', minWidth: 200, maxWidth: 200},
+      {headerName: 'Teamnaam', valueGetter: this.formatTeamnaam, minWidth: 200, maxWidth: 200},
       {headerName: 'Totaal', valueGetter: this.formatTotaalpunten, minWidth: 100, maxWidth: 100},
       {headerName: 'Etappes', field: 'totalStagePoints', minWidth: 100, maxWidth: 100},
       {
@@ -89,26 +91,24 @@ export class ParticipanttableComponent implements OnInit {
     this.participantstable$ = this.store.pipe(select(getParticipanttable));
     this.lastUpdated$ = this.store.pipe(select(getLastUpdated));
 
-    this.lastUpdated$.subscribe(lastupdated => {
+    this.lastUpdated$.pipe(takeUntil(this.unsubscribe)).subscribe(lastupdated => {
       if (lastupdated) {
         this.lastUpdated = moment(lastupdated.lastUpdated).fromNow();
       }
     });
     this.gridOptions = <GridOptions>{
-      context: {parentComponent: this},
-      columnDefs: this.agColumns,
-
-      onGridReady: () => {
-        this.gridOptions.api.sizeColumnsToFit();
+      defaultColDef: {
+        sortable: true,
+        resizable: true
       },
-      enableSorting: true,
+      context: {parentComponent: this},
+      columnDefs: this.agColumns
     };
   }
-
   onGridReady(params) {
     this.gridApi = params.api;
     this.gridColumnApi = params.columnApi;
-    this.store.select(getTour).subscribe(tour => {
+    this.store.pipe(select(getTour)).pipe(takeUntil(this.unsubscribe)).subscribe(tour => {
       this.tour = tour;
     });
 
@@ -126,7 +126,7 @@ export class ParticipanttableComponent implements OnInit {
   }
 
   formatPosition(params): string {
-    const deltaPositie = params.data.previousPosition - params.data.position;
+    const deltaPositie = params.data.previousPosition ? params.data.previousPosition - params.data.position : 0;
     if (deltaPositie > 0) {
       return params.data.position + ' (+' + deltaPositie + ')';
     } else {
@@ -141,10 +141,21 @@ export class ParticipanttableComponent implements OnInit {
     return params.data.totalTourPoints + params.data.totalMountainPoints + params.data.totalPointsPoints + params.data.totalYouthPoints;
   }
 
+
+  formatTeamnaam(params): string {
+    return params.data.teamName ? params.data.teamName : params.data.displayName;
+  }
+
   formatTotaalpunten(params): string {
     const addendum: string =
-      (params.data.deltaTotalStagePoints > 0) ? ' (+' + params.data.deltaTotalStagePoints + ')' :
-        (params.data.deltaTotalStagePoints === 0) ? '' : ' (' + params.data.deltaTotalStagePoints + ')';
+      (params.data.deltaTotalStagePoints && params.data.deltaTotalStagePoints > 0) ? ' (+' + params.data.deltaTotalStagePoints + ')' :
+        (!params.data.deltaTotalStagePoints || params.data.deltaTotalStagePoints === 0) ? '' :
+          ' (' + params.data.deltaTotalStagePoints + ')';
     return params.data.totalPoints + addendum;
+  }
+
+  ngOnDestroy() {
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
   }
 }
