@@ -1,20 +1,20 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {TourriderDetailComponent} from './tourrider-detail.component';
 
-import { TourriderdetaildialogComponent } from './tourriderdetaildialog.component';
 
 describe('TourriderdetaildialogComponent', () => {
-  let component: TourriderdetaildialogComponent;
-  let fixture: ComponentFixture<TourriderdetaildialogComponent>;
+  let component: TourriderDetailComponent;
+  let fixture: ComponentFixture<TourriderDetailComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ TourriderdetaildialogComponent ]
+      declarations: [ TourriderDetailComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(TourriderdetaildialogComponent);
+    fixture = TestBed.createComponent(TourriderDetailComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
