@@ -15,10 +15,21 @@ export class EtappeEffects {
   fetchEtappeList$ = this.actions$
     .pipe(ofType<etappe.FetchEtappeList>(etappe.FETCH_ETAPPELIST),
       switchMap(action => {
-      return this.etappeService
-        .getEtappes(action.payload)
-        .pipe(switchMap(etappeResponse =>
-            of(new etappe.FetchEtappeListSuccess(etappeResponse))),
-          catchError(err => of(new etappe.FetchEtappeListFailure(err))));
-    }));
+        return this.etappeService
+          .getEtappes(action.payload)
+          .pipe(switchMap(etappeResponse =>
+              of(new etappe.FetchEtappeListSuccess(etappeResponse))),
+            catchError(err => of(new etappe.FetchEtappeListFailure(err))));
+      }));
+
+  @Effect()
+  fetchLatestEtappe$ = this.actions$
+    .pipe(ofType<etappe.FetchLatestEtappe>(etappe.FETCH_LATESTETAPPE),
+      switchMap(action => {
+        return this.etappeService
+          .getLatestDrivenEtappe(action.payload)
+          .pipe(switchMap(etappeResponse =>
+              of(new etappe.FetchLatestEtappeSuccess(etappeResponse))),
+            catchError(err => of(new etappe.FetchLatestEtappeFailure(err))));
+      }));
 }
