@@ -10,12 +10,7 @@ export interface ParticipanttableState {
 }
 
 const initalparticipanttableState: ParticipanttableState = {
-  participanttable: [{
-    id: null,
-    displayName: null,
-    predictions: [],
-    totalPoints: null,
-  }],
+  participanttable: [],
   lastUpdated: null,
   error: undefined,
   inProgress: false,
@@ -72,6 +67,7 @@ export function participanttableReducer(state = initalparticipanttableState, act
 export const getparticipanttableState = createFeatureSelector<ParticipanttableState>('participanttable');
 export const getParticipanttable = createSelector(getparticipanttableState, (state: ParticipanttableState) => state.participanttable);
 export const getNumberOne = createSelector(getparticipanttableState, (state: ParticipanttableState) => state.participanttable[0]);
+export const getTopX = x => createSelector(getparticipanttableState, (state: ParticipanttableState) => state.participanttable.slice(0, x));
 export const getLastUpdated = createSelector(getparticipanttableState, (state: ParticipanttableState) => state.lastUpdated);
 export const getParticipantPredictions = id =>
   createSelector(getparticipanttableState, (state: ParticipanttableState) =>
