@@ -1,9 +1,11 @@
 import {Injectable} from '@angular/core';
 import {Actions, Effect, ofType} from '@ngrx/effects';
 import * as tour from './tour.actions';
+import * as etappe from '../etappe/etappe.actions';
+import * as tourrider from '../tourriders/tourrider.actions';
 
 import {TourService} from '../../services/tour.service';
-import {of} from 'rxjs';
+import {from, of} from 'rxjs';
 import {catchError, switchMap} from 'rxjs/operators';
 
 @Injectable()
@@ -35,6 +37,15 @@ export class TourEffects {
             catchError(err => of(new tour.FetchTourFailure(err))));
       }));
 
+  @Effect()
+  fetchTourSuccess$ = this.actions$
+    .pipe(ofType<tour.FetchTourSuccess>(tour.FETCH_TOUR_SUCCESS),
+      switchMap(action =>
+        from([new etappe.FetchEtappeList(action.payload.id),
+          new etappe.FetchLatestEtappe(action.payload.id),
+          new tourrider.FetchTourriderList(action.payload.id)])),
+      catchError(err => of(new tour.FetchTourFailure(err))));
+
 
   @Effect()
   fetchTourList$ = this.actions$
@@ -44,7 +55,7 @@ export class TourEffects {
         return this.tourService
           .getTourlist()
           .pipe(switchMap(tourResponse =>
-            of(new tour.FetchTourListSuccess(tourResponse))
+              of(new tour.FetchTourListSuccess(tourResponse))
             ),
             catchError(err => of(new tour.FetchTourListFailure(err))));
       }));
